requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-calendar>=0.8.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0

# AI and Language Models
openai>=1.0.0
langchain>=0.0.350
langchain-openai>=0.0.5

# Microsoft Graph API integration
requests>=2.31.0
msal>=1.24.0

# Date and time utilities
python-dateutil>=2.8.2

# Environment variables
python-dotenv>=1.0.0

# HTTP requests and API calls
httpx>=0.25.0

# Type hints
typing-extensions>=4.7.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Additional utilities
pydantic>=2.0.0